+++ lizmap/forms/filter_by_form_edition.py
@@ -1,7 +1,9 @@
 """Dialog for filter by form."""
 
 from qgis.core import QgsMapLayerProxyModel, QgsProject
+from qgis.PyQt.QtGui import QIcon
 
+from lizmap import LwcVersions
 from lizmap.definitions.filter_by_form import FilterByFormDefinitions
 from lizmap.forms.base_edition_dialog import BaseEditionDialog
 from lizmap.qgis_plugin_tools.tools.i18n import tr
@@ -26,8 +28,8 @@ class FilterByFormEditionDialog(BaseEditionDialog, CLASS):
         self.config.add_layer_widget('title', self.title)
         self.config.add_layer_widget('type', self.type)
         self.config.add_layer_widget('field', self.field)
-        self.config.add_layer_widget('min_date', self.min_date)
-        self.config.add_layer_widget('max_date', self.max_date)
+        self.config.add_layer_widget('start_field', self.start_field)
+        self.config.add_layer_widget('end_field', self.end_field)
         self.config.add_layer_widget('format', self.filter_format)
         self.config.add_layer_widget('splitter', self.splitter)
 
@@ -35,33 +37,58 @@ class FilterByFormEditionDialog(BaseEditionDialog, CLASS):
         self.config.add_layer_label('title', self.label_title)
         self.config.add_layer_label('type', self.label_type)
         self.config.add_layer_label('field', self.label_field)
-        self.config.add_layer_label('min_date', self.label_min_date)
-        self.config.add_layer_label('max_date', self.label_max_date)
+        self.config.add_layer_label('start_field', self.label_start_field)
+        self.config.add_layer_label('end_field', self.label_end_field)
         self.config.add_layer_label('format', self.label_format)
         self.config.add_layer_label('splitter', self.label_splitter)
 
-        self.type.addItem(tr('Text'), 'text')
-        self.type.addItem(tr('Unique values'), 'uniquevalues')
-        self.type.addItem(tr('Numeric'), 'numeric')
-        self.type.addItem(tr('Date'), 'date')
+        # TODO this shouldn't be here but in the definitions package
+        self.type.addItem(
+            QIcon(":images/themes/default/mIconFieldText.svg"),
+            tr('Text'),
+            'text'
+        )
+        self.type.addItem(
+            QIcon(":images/themes/default/algorithms/mAlgorithmUniqueValues.svg"),
+            tr('Unique values'),
+            'uniquevalues'
+        )
+        self.type.addItem(
+            QIcon(":images/themes/default/mIconFieldInteger.svg"),
+            tr('Numeric'),
+            'numeric',
+        )
+        self.type.addItem(
+            QIcon(":images/themes/default/mIconFieldDateTime.svg"),
+            tr('Date'),
+            'date',
+        )
 
-        self.filter_format.addItem(tr('Checkboxes'), 'checkboxes')
-        self.filter_format.addItem(tr('Combobox'), 'select')
+        self.filter_format.addItem(
+            QIcon(":images/themes/default/mIconDeselected.svg"),
+            tr('Checkboxes'),
+            'checkboxes'
+        )
+        self.filter_format.addItem(
+            QIcon(":images/themes/default/mIconDropDownMenu.svg"),
+            tr('Combobox'),
+            'select'
+        )
 
         self.layer.setFilters(QgsMapLayerProxyModel.VectorLayer)
 
         self.layer.layerChanged.connect(self.check_layer_wfs)
         self.layer.layerChanged.connect(self.field.setLayer)
-        self.layer.layerChanged.connect(self.min_date.setLayer)
-        self.layer.layerChanged.connect(self.max_date.setLayer)
+        self.layer.layerChanged.connect(self.start_field.setLayer)
+        self.layer.layerChanged.connect(self.end_field.setLayer)
 
         self.field.setLayer(self.layer.currentLayer())
-        self.min_date.setLayer(self.layer.currentLayer())
-        self.max_date.setLayer(self.layer.currentLayer())
+        self.start_field.setLayer(self.layer.currentLayer())
+        self.end_field.setLayer(self.layer.currentLayer())
 
-        self.field.setAllowEmptyFieldName(False)
-        self.min_date.setAllowEmptyFieldName(False)
-        self.max_date.setAllowEmptyFieldName(True)
+        self.lwc_versions[LwcVersions.Lizmap_3_7] = [
+            # self.label_end_field, this one, it depends on the type of filter : date versus numeric
+        ]
 
         block_list = []
         for layer in QgsProject().instance().mapLayers().values():
@@ -85,51 +112,74 @@ class FilterByFormEditionDialog(BaseEditionDialog, CLASS):
         self.show_error(not_in_wfs)
 
     def update_visibility(self):
-        """Show/Hide fields depending of chosen type."""
+        """Show/Hide fields depending on chosen type."""
         index = self.type.currentIndex()
         data = self.type.itemData(index)
 
-        if data == 'date':
-            self.min_date.setVisible(True)
-            self.min_date.setAllowEmptyFieldName(False)
-            self.max_date.setVisible(True)
-            self.label_min_date.setVisible(True)
-            self.label_max_date.setVisible(True)
-        else:
-            self.min_date.setVisible(False)
-            self.min_date.setAllowEmptyFieldName(True)
-            self.min_date.setField('')
-            self.max_date.setVisible(False)
-            self.max_date.setField('')
-            self.label_min_date.setVisible(False)
-            self.label_max_date.setVisible(False)
+        # Make all not common fields not visible by default
+        # And enable them one per one
+        # Field
+        self.label_field.setVisible(False)
+        self.field.setVisible(False)
+        self.field.setAllowEmptyFieldName(True)
 
-        if data == 'uniquevalues':
+        # Start field
+        self.label_start_field.setVisible(False)
+        self.start_field.setVisible(False)
+        self.start_field.setAllowEmptyFieldName(True)
+
+        # End field
+        self.label_end_field.setVisible(False)
+        self.end_field.setVisible(False)
+        self.end_field.setAllowEmptyFieldName(True)
+
+        # Format
+        self.filter_format.setVisible(False)
+        self.label_format.setVisible(False)
+
+        # Splitter
+        self.splitter.setVisible(False)
+        self.label_splitter.setVisible(False)
+
+        if data == 'date':
+            self.start_field.setVisible(True)
+            self.start_field.setAllowEmptyFieldName(False)
+            self.end_field.setVisible(True)
+            self.label_start_field.setVisible(True)
+            self.label_end_field.setVisible(True)
+            if self.label_end_field in self.lwc_versions[LwcVersions.Lizmap_3_7]:
+                self.lwc_versions[LwcVersions.Lizmap_3_7].remove(self.label_end_field)
+        elif data == 'uniquevalues':
+            self.label_field.setVisible(True)
+            self.field.setVisible(True)
+            self.field.setAllowEmptyFieldName(False)
             self.filter_format.setVisible(True)
             index = self.filter_format.findData('')
             self.filter_format.removeItem(index)
             self.splitter.setVisible(True)
             self.label_format.setVisible(True)
             self.label_splitter.setVisible(True)
-        else:
-            self.filter_format.setVisible(False)
+        elif data == 'numeric':
+            self.label_start_field.setVisible(True)
+            self.label_end_field.setVisible(True)
+            self.start_field.setVisible(True)
+            self.start_field.setAllowEmptyFieldName(False)
+            self.end_field.setVisible(True)
+            if self.label_end_field not in self.lwc_versions[LwcVersions.Lizmap_3_7]:
+                self.lwc_versions[LwcVersions.Lizmap_3_7].append(self.label_end_field)
+        elif data == 'text':
+            self.label_field.setVisible(True)
+            self.field.setVisible(True)
+            self.field.setAllowEmptyFieldName(False)
             self.filter_format.addItem('', '')
             index = self.filter_format.findData('')
             self.filter_format.setCurrentIndex(index)
-            self.splitter.setVisible(False)
             self.splitter.setText('')
-            self.label_format.setVisible(False)
-            self.label_splitter.setVisible(False)
-
-        if data in ['text', 'uniquevalues', 'numeric']:
-            self.field.setVisible(True)
-            self.field.setAllowEmptyFieldName(False)
-            self.label_field.setVisible(True)
         else:
-            self.field.setVisible(False)
-            self.field.setAllowEmptyFieldName(True)
-            self.field.setField('')
-            self.label_field.setVisible(False)
+            raise Exception('Unknown type')
+
+        # Let's repaint colors on widgets because of the numeric versus date type
+        self.version_lwc()
 
     def validate(self) -> str:
         upstream = super().validate()
@@ -144,11 +194,12 @@ class FilterByFormEditionDialog(BaseEditionDialog, CLASS):
         index = self.type.currentIndex()
         data = self.type.itemData(index)
 
-        if data == 'date':
-            if not self.min_date.currentField():
-                return tr('Field min date is mandatory.')
-        elif data in ['text', 'uniquevalues', 'numeric']:
+        field_required = tr('Field is mandatory.')
+        if data in ('numeric', 'date'):
+            if not self.start_field.currentField():
+                return field_required
+        elif data in ('text', 'uniquevalues'):
             if not self.field.currentField():
-                return tr('Field is mandatory.')
+                return field_required
         else:
             raise Exception('Unknown option')