+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Add new "maximum field" for Lizmap Web Client ≥ 3.7 about the form filter with numeric values
 * Remove notes about the Lizmap QGIS server plugin
 
 ## 3.9.5 - 2022-11-29